% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/upc.R
\name{upc}
\alias{upc}
\title{UPC model}
\usage{
upc(fig_width = 4, fig_height = 2.5, fig_crop = TRUE, toc = FALSE,
  toc_depth = 2, number_sections = TRUE, highlight = "default",
  fig_caption = FALSE, keep_tex = FALSE, includes = NULL,
  pandoc_args = NULL)
}
\description{
Model per generar plantilles de la UPC
}

