% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/upcTemplate.R
\docType{package}
\name{upcTemplate}
\alias{upcTemplate}
\alias{upcTemplate-package}
\title{Plantilla per projectes de la UPC}
\description{
Plantilla generada amb rmarkdown per projectes de la UPC.
}

